# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Zentrivex Crypto Investment Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterInput {
  email: string;
  /** @minLength 8 */
  password: string;
  firstName: string;
  lastName: string;
  /** @nullable */
  phone?: string | null;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  user: 'user',
  admin: 'admin',
} as const;

export type UserKycStatus = typeof UserKycStatus[keyof typeof UserKycStatus];


export const UserKycStatus = {
  none: 'none',
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface User {
  id: number;
  email: string;
  firstName: string;
  lastName: string;
  /** @nullable */
  phone?: string | null;
  role: UserRole;
  balance: number;
  kycStatus: UserKycStatus;
  isActive?: boolean;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export type UserUpdateRole = typeof UserUpdateRole[keyof typeof UserUpdateRole];


export const UserUpdateRole = {
  user: 'user',
  admin: 'admin',
} as const;

export interface UserUpdate {
  isActive?: boolean;
  balance?: number;
  role?: UserUpdateRole;
}

export interface Plan {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  minAmount: number;
  maxAmount: number;
  roiPercent: number;
  durationDays: number;
  isActive: boolean;
  totalReturn?: number;
  createdAt?: string;
}

export interface PlanInput {
  name: string;
  description?: string;
  minAmount: number;
  maxAmount: number;
  roiPercent: number;
  durationDays: number;
  isActive?: boolean;
}

export interface PlanUpdate {
  name?: string;
  description?: string;
  minAmount?: number;
  maxAmount?: number;
  roiPercent?: number;
  durationDays?: number;
  isActive?: boolean;
}

export type InvestmentStatus = typeof InvestmentStatus[keyof typeof InvestmentStatus];


export const InvestmentStatus = {
  active: 'active',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface Investment {
  id: number;
  userId: number;
  planId: number;
  amount: number;
  status: InvestmentStatus;
  profit?: number;
  startDate: string;
  endDate: string;
  plan?: Plan;
  createdAt: string;
}

export interface InvestmentInput {
  planId: number;
  amount: number;
}

export type DepositStatus = typeof DepositStatus[keyof typeof DepositStatus];


export const DepositStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface Deposit {
  id: number;
  userId: number;
  amount: number;
  currency: string;
  walletAddress: string;
  /** @nullable */
  txHash?: string | null;
  /** @nullable */
  proofImage?: string | null;
  status: DepositStatus;
  /** @nullable */
  rejectionReason?: string | null;
  user?: User;
  createdAt: string;
  /** @nullable */
  approvedAt?: string | null;
}

export interface DepositInput {
  amount: number;
  currency: string;
  walletAddress: string;
  txHash?: string;
  proofImage?: string;
}

export type WithdrawalStatus = typeof WithdrawalStatus[keyof typeof WithdrawalStatus];


export const WithdrawalStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface Withdrawal {
  id: number;
  userId: number;
  amount: number;
  walletAddress: string;
  /** @nullable */
  network?: string | null;
  status: WithdrawalStatus;
  /** @nullable */
  rejectionReason?: string | null;
  user?: User;
  createdAt: string;
  /** @nullable */
  approvedAt?: string | null;
}

export interface WithdrawalInput {
  amount: number;
  walletAddress: string;
  network?: string;
}

export type KycDocumentType = typeof KycDocumentType[keyof typeof KycDocumentType];


export const KycDocumentType = {
  passport: 'passport',
  drivers_license: 'drivers_license',
  national_id: 'national_id',
} as const;

export type KycStatus = typeof KycStatus[keyof typeof KycStatus];


export const KycStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface Kyc {
  id: number;
  userId: number;
  documentType: KycDocumentType;
  /** @nullable */
  frontImage?: string | null;
  /** @nullable */
  backImage?: string | null;
  /** @nullable */
  selfieImage?: string | null;
  status: KycStatus;
  /** @nullable */
  rejectionReason?: string | null;
  user?: User;
  createdAt: string;
  /** @nullable */
  reviewedAt?: string | null;
}

export type KycInputDocumentType = typeof KycInputDocumentType[keyof typeof KycInputDocumentType];


export const KycInputDocumentType = {
  passport: 'passport',
  drivers_license: 'drivers_license',
  national_id: 'national_id',
} as const;

export interface KycInput {
  documentType: KycInputDocumentType;
  frontImage?: string;
  backImage?: string;
  selfieImage?: string;
}

export interface RejectInput {
  reason: string;
}

export interface DashboardSummary {
  balance: number;
  totalInvested: number;
  totalProfit: number;
  totalDeposited: number;
  totalWithdrawn: number;
  activeInvestments: number;
  pendingDeposits?: number;
  pendingWithdrawals?: number;
  kycStatus?: string;
}

export interface AdminDashboardSummary {
  totalUsers: number;
  totalDeposits: number;
  totalWithdrawals: number;
  totalInvestments: number;
  pendingDeposits: number;
  pendingWithdrawals: number;
  pendingKyc: number;
  platformBalance: number;
  recentDeposits?: Deposit[];
  recentWithdrawals?: Withdrawal[];
}

export type TransactionType = typeof TransactionType[keyof typeof TransactionType];


export const TransactionType = {
  deposit: 'deposit',
  withdrawal: 'withdrawal',
  investment: 'investment',
  profit: 'profit',
  referral: 'referral',
} as const;

export type TransactionStatus = typeof TransactionStatus[keyof typeof TransactionStatus];


export const TransactionStatus = {
  pending: 'pending',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface Transaction {
  id: number;
  userId: number;
  type: TransactionType;
  amount: number;
  status: TransactionStatus;
  /** @nullable */
  description?: string | null;
  createdAt: string;
}