# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Zentrivex Crypto Investment Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
...
  /investments:
    get:
      operationId: listInvestments
      tags: [investments]
      summary: List user investments
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Investment"
    post:
      operationId: createInvestment
      tags: [investments]
      summary: Start investment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvestmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Investment"
  /investments/{id}:
    get:
      operationId: getInvestment
      tags: [investments]
      summary: Get investment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Investment"
  /deposits:
    get:
      operationId: listDeposits
      tags: [deposits]
      summary: List user deposits
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deposit"
    post:
      operationId: createDeposit
      tags: [deposits]
      summary: Submit deposit request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deposit"
...
  /admin/users/{id}:
    get:
      operationId: getUser
      tags: [admin]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [admin]
      summary: Update user (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /admin/deposits/{id}/approve:
    patch:
      operationId: approveDeposit
      tags: [admin]
      summary: Approve deposit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deposit"
  /admin/deposits/{id}/reject:
    patch:
      operationId: rejectDeposit
      tags: [admin]
      summary: Reject deposit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deposit"
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [email, password, firstName, lastName]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: ["string", "null"]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, email, firstName, lastName, role, balance, kycStatus, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, admin]
        balance:
          type: number
        kycStatus:
          type: string
          enum: [none, pending, approved, rejected]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    UserUpdate:
      type: object
      properties:
        isActive:
          type: boolean
        balance:
          type: number
        role:
          type: string
          enum: [user, admin]
    Plan:
      type: object
      required: [id, name, minAmount, maxAmount, roiPercent, durationDays, isActive]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        minAmount:
          type: number
        maxAmount:
          type: number
        roiPercent:
          type: number
        durationDays:
          type: integer
        isActive:
          type: boolean
        totalReturn:
          type: number
        createdAt:
          type: string
          format: date-time
...
    Investment:
      type: object
      required: [id, userId, planId, amount, status, startDate, endDate, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        planId:
          type: integer
        amount:
          type: number
        status:
          type: string
          enum: [active, completed, cancelled]
        profit:
          type: number
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        plan:
          $ref: "#/components/schemas/Plan"
        createdAt:
          type: string
          format: date-time
    InvestmentInput:
      type: object
      required: [planId, amount]
      properties:
        planId:
          type: integer
        amount:
          type: number
    Deposit:
      type: object
      required: [id, userId, amount, currency, walletAddress, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        amount:
          type: number
        currency:
          type: string
        walletAddress:
          type: string
        txHash:
          type: ["string", "null"]
        proofImage:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected]
        rejectionReason:
          type: ["string", "null"]
        user:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
          format: date-time
        approvedAt:
          type: ["string", "null"]
    DepositInput:
      type: object
      required: [amount, currency, walletAddress]
      properties:
        amount:
          type: number
        currency:
          type: string
        walletAddress:
          type: string
        txHash:
          type: string
        proofImage:
          type: string
...
    RejectInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
...