# lib/api-zod/src/generated/types/registerInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Zentrivex Crypto Investment Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterInput {
  email: string;
  /** @minLength 8 */
  password: string;
  firstName: string;
  lastName: string;
  /** @nullable */
  phone?: string | null;
}