# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Zentrivex Crypto Investment Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserKycStatus } from './userKycStatus';
import type { UserRole } from './userRole';

export interface User {
  id: number;
  email: string;
  firstName: string;
  lastName: string;
  /** @nullable */
  phone?: string | null;
  role: UserRole;
  balance: number;
  kycStatus: UserKycStatus;
  isActive?: boolean;
  createdAt: Date;
}