# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Zentrivex Crypto Investment Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserUpdateRole } from './userUpdateRole';

export interface UserUpdate {
  isActive?: boolean;
  balance?: number;
  role?: UserUpdateRole;
}